model_pipeline/requirements.txt
# =============================================================================
# Model Pipeline - Requirements
# Python 3.11 Compatible
# =============================================================================

# -----------------------------------------------------------------------------
# Core ML & Data Processing
# -----------------------------------------------------------------------------
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<2.3.0
scipy>=1.11.0,<1.14.0
scikit-learn>=1.3.0,<1.6.0

# -----------------------------------------------------------------------------
# XGBoost
# -----------------------------------------------------------------------------
xgboost>=2.0.0,<2.2.0

# -----------------------------------------------------------------------------
# MLflow
# -----------------------------------------------------------------------------
mlflow>=2.15.0,<2.20.0
boto3>=1.34.0,<1.42.0
cloudpickle>=3.0.0,<3.2.0

# -----------------------------------------------------------------------------
# Model Explainability
# -----------------------------------------------------------------------------
shap>=0.44.0,<0.46.0

# -----------------------------------------------------------------------------
# Configuration & Logging
# -----------------------------------------------------------------------------
PyYAML>=6.0.0,<6.1.0
loguru>=0.7.0,<0.8.0
python-dotenv>=1.0.0,<1.1.0

# -----------------------------------------------------------------------------
# Data Validation
# -----------------------------------------------------------------------------
pydantic>=2.5.0,<2.10.0

# -----------------------------------------------------------------------------
# Testing
# -----------------------------------------------------------------------------
pytest>=7.4.0,<8.1.0
pytest-cov>=4.1.0,<4.2.0

# -----------------------------------------------------------------------------
# Visualization
# -----------------------------------------------------------------------------
matplotlib>=3.8.0,<3.10.0
seaborn>=0.13.0,<0.14.0

# -----------------------------------------------------------------------------
# Utils
# -----------------------------------------------------------------------------
tqdm>=4.66.0,<4.67.0
joblib>=1.3.0,<1.5.0
